KNOX-2861 upgrade CM API version

+++ gateway-discovery-cm/src/main/java/org/apache/knox/gateway/topology/discovery/cm/monitor/PollingConfigurationAnalyzer.java
@@ -48,6 +48,7 @@ import org.apache.knox.gateway.topology.simple.SimpleDescriptorFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
@@ -462,7 +463,8 @@ public class PollingConfigurationAnalyzer implements Runnable {
 
     try {
       // giving 'null' as maximum result size results in fetching all events from CM within the given time interval
-      ApiEventQueryResult eventsResult = (new EventsResourceApi(client)).readEvents(null, queryString, 0);
+      ApiEventQueryResult eventsResult = (new EventsResourceApi(client)).readEvents(null, queryString,
+          BigDecimal.valueOf(0));
       events.addAll(eventsResult.getItems());
     } catch (ApiException e) {
       log.clouderaManagerEventsAPIError(e);
